requirements.txt
Pillow
numpy
pyrr
random
glfw
PyOpenGL
matplotlib